requirements.txt
gradio
librosa
scikit-learn
xgboost
numpy
matplotlib
soundfile
joblib
pillow
scipy
fpdf2
soxr# updated
